+++ code/edgemedia/src/main/java/com/adobe/marketing/mobile/edge/media/internal/MediaXDMEventHelper.kt
@@ -217,10 +217,10 @@ internal class MediaXDMEventHelper {
             }
 
             return XDMQoeDataDetails(
-                qoeInfo.bitrate.toLong(),
-                qoeInfo.droppedFrames.toLong(),
-                qoeInfo.fps.toLong(),
-                qoeInfo.startupTime.toLong()
+                qoeInfo.bitrate,
+                qoeInfo.droppedFrames,
+                qoeInfo.fps,
+                qoeInfo.startupTime
             )
         }
